pyproject.toml
[project]
name = "easy-ai18n"
version = "0.0.5"
description = "Simple and Elegant Python3 Internationalization (i18n) Tool"
requires-python = ">=3.12"
dependencies = [
    "aioresult==1.2",
    "anyio>=4.9.0",
    "googletrans==4.0.2",
    "instructor>=1.7.9",
    "loguru>=0.7.3",
    "openai==1.75.0",
    "pytest==8.3.5",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "tenacity==9.1.2",
    "tqdm>=4.67.1",
]
[project.urls]
Repository = "https://github.com/z-mio/easy-ai18n"
Issues = "https://github.com/z-mio/easy-ai18n/issues"